Fixed issue #26: slide numbering is just plain wrong.
Previously we were incrementing the slide count:

a) as soon as we visited a slide, and
b) regardless of whether or not this slide was going to be rendered as
   a slide or a div.

After incrementing, however, we'd often call depart_slide to close the
previous slide. To compensate for (a), _add_slide_number was called
with (count - 1) when departing a slide.

This commit simplifies the logic: we increment right before we call
starttag, and only when we're creating a slide. This means we also
don't need to worry about subtracting 1 to get the actual slide
number.