Always reread documents that transform the doctree (#51)
The Sphinx doctree is pickled to disk, and by default that pickle is
shared between builders. The ``nextslide`` and slide conditional
directives transform the doctree before it's pickled, which can lead to
unexpected results when building both HTML and slides from the same
source.

If a document contains one of these nodes, we tell the environment to
note it should be reread next time, regardless of whether it has been
changed.