Use `load` instead of `resolve` for theme entrypoints (#94, #57).
The `load` method exists in older versions of setuptools, and has the
side-effect of `require`-ing the module needed for the entry point.
`resolve` is newer, and only operates in the current working set.
There shouldn't be a downside to using `load` since all themes are
currently provided by the `hieroglyph` package.